also split out EnumVariantValueAssigner
